+++ openquake/hazardlib/tests/sourceconverter_test.py
@@ -21,6 +21,7 @@ import unittest.mock
 import numpy
 from openquake.baselib import hdf5
 from openquake.hazardlib import nrml
+from openquake.hazardlib.geo import Point
 from openquake.hazardlib.sourceconverter import update_source_model, \
     SourceConverter
 
@@ -277,3 +278,37 @@ class SourceGroupHDF5TestCase(unittest.TestCase):
             f['grp'] = grp
         with hdf5.File(f.path, 'r') as f:
             print(f['grp'])
+
+
+class MultiFaultSourceModelTestCase(unittest.TestCase):
+    """ Tests reading a multi fault model """
+
+    def test_load_mfs(self):
+        datadir = os.path.join(os.path.dirname(__file__), 'data', 'sections')
+        sec_xml = os.path.join(datadir, 'sections_kite.xml')
+        src_xml = os.path.join(datadir, 'sources.xml')
+        conv = SourceConverter()
+        sec = nrml.to_python(sec_xml, conv).sections
+        expected = [Point(11, 45, 0), Point(11, 45.5, 10)]
+        # Check geometry info
+        self.assertEqual(expected[0], sec[1].surface.profiles[0].points[0])
+        self.assertEqual(expected[1], sec[1].surface.profiles[0].points[1])
+        self.assertEqual(sec[1].sec_id, 's2')
+        ssm = nrml.to_python(src_xml, conv)
+        self.assertIsInstance(ssm, nrml.SourceModel)
+
+        rups = []
+        ssm[0][0].sections = sec  # fix sections
+        ssm[0][0].create_inverted_index()
+        for rup in ssm[0][0].iter_ruptures():
+            rups.append(rup)
+
+        # Check data for the second rupture
+        msg = 'Rake for rupture #0 is wrong'
+        self.assertEqual(rups[0].rake, 90.0, msg)
+        # Check data for the second rupture
+        msg = 'Rake for rupture #1 is wrong'
+        self.assertEqual(rups[1].rake, -90.0, msg)
+        # Check mfd
+        expected = numpy.array([0.9, 0.1])
+        numpy.testing.assert_almost_equal(rups[1].probs_occur, expected)
